Began mapper class and adjusted location and IIngredients to match database entries. Also changed SizeModdifier method to use int instead of double

+++ project 1/PizzaStoreApplication/PizzaStoreApplicationLibrary/Models/IIngredients.cs
@@ -6,16 +6,16 @@ namespace PizzaStoreApplicationLibrary
 {
     public interface IIngredients
     {
-        double Dough { get; set; }
+        int Dough { get; set; }
 
-        double Cheese { get; set; }
+        int Cheese { get; set; }
 
-        double Sauce { get; set; }
+        int Sauce { get; set; }
 
-        double Pepperoni { get; set; }
+        int Pepperoni { get; set; }
 
-        double PeppersAndOnions { get; set; }
+        int PeppersAndOnions { get; set; }
 
-        double HamAndMeatball { get; set; }
+        int HamAndMeatball { get; set; }
     }
 }
